Some incremental work towards consumption-connector-service
