+++ libs/services/designer-client-services/src/lib/standard/connector.ts
@@ -1,3 +1,61 @@
 import { BaseConnectorService } from '../base';
+import type { ListDynamicValue } from '../connector';
 
-export class StandardConnectorService extends BaseConnectorService {}
+export class StandardConnectorService extends BaseConnectorService {
+  async getListDynamicValues(
+    connectionId: string | undefined,
+    connectorId: string,
+    operationId: string,
+    _parameterAlias: string | undefined,
+    parameters: Record<string, any>,
+    dynamicState: any
+  ): Promise<ListDynamicValue[]> {
+    const { baseUrl, apiVersion, getConfiguration, httpClient } = this.options;
+    const { operationId: dynamicOperation } = dynamicState;
+
+    const invokeParameters = this._getInvokeParameters(parameters, dynamicState);
+    const configuration = await getConfiguration(connectionId ?? '');
+
+    if (this._isClientSupportedOperation(connectorId, operationId)) {
+      return this.options.valuesClient[dynamicOperation]({ parameters: invokeParameters, configuration });
+    }
+
+    const uri = `${baseUrl}/operationGroups/${connectorId.split('/').slice(-1)}/operations/${dynamicOperation}/dynamicInvoke`;
+    const response = await httpClient.post({
+      uri,
+      queryParameters: { 'api-version': apiVersion },
+      content: { parameters: invokeParameters, configuration },
+    });
+    return this._getResponseFromDynamicApi(response, uri);
+  }
+
+  async getDynamicSchema(
+    connectionId: string | undefined,
+    connectorId: string,
+    operationId: string,
+    _parameterAlias: string | undefined,
+    parameters: Record<string, any>,
+    dynamicState: any
+  ): Promise<OpenAPIV2.SchemaObject> {
+    const { baseUrl, apiVersion, getConfiguration, httpClient } = this.options;
+    const {
+      extension: { operationId: dynamicOperation },
+      isInput,
+    } = dynamicState;
+
+    const invokeParameters = this._getInvokeParameters(parameters, dynamicState);
+    const configuration = await getConfiguration(connectionId ?? '');
+
+    if (this._isClientSupportedOperation(connectorId, operationId)) {
+      return this.options.schemaClient[dynamicOperation]({ parameters: invokeParameters, isInput, configuration });
+    }
+
+    const uri = `${baseUrl}/operationGroups/${connectorId.split('/').slice(-1)}/operations/${dynamicOperation}/dynamicInvoke`;
+    const response = await httpClient.post({
+      uri,
+      queryParameters: { 'api-version': apiVersion },
+      content: { parameters: invokeParameters, configuration },
+    });
+    return this._getResponseFromDynamicApi(response, uri);
+  }
+}
